fix: fix player hints #99
