+++ shared/helper.js
@@ -1,3 +1,6 @@
+import { deepCopy } from "./helperFunctions/deepCopy";
+import { popChoice } from "./helperFunctions/popChoice";
+
 export function isMultiPlayer(playerCount) {
   if (playerCount > 1) {
     return true;
@@ -159,4 +162,85 @@ export function getNeighborPlayers(player, game) {
   });
 
   return neighbors;
+}
+
+export function getHints(player, round, game) {
+  const { treatment } = game;
+  const task = round.get("task");
+  const possibleHints = task.question.hints;
+
+  if (!possibleHints || possibleHints.length === 0) {
+    return null;
+  }
+
+  //To allow for random hints we need to ensure:
+  // - the hint won't change every time the component updates
+  // - the hint will have the possibility to be different for every participant
+  // SOLUTION: Set the hints to the player
+  // The player will get a "hints" attribute on gameInit which is an empty object
+  // The round will have an index set according to the round number it is (0 to nRound)
+  // As the rounds go, the "hints" of the player will be populated, where each key is the index of the round
+  // BONUS: This allows us to track which hints player got at which round.
+
+  // If the hints are already set you want to get those instead of resetting them...
+  let hints;
+  if (typeof player.get("hints")[round.get("index")] != "undefined") {
+    hints = player.get("hints")[round.get("index")];
+  } else {
+    //If they are not set, you want to set them...
+    hints = [];
+
+    // If there is a configuration set in the treatment.hint, use it to select the hints
+    if (treatment.hints) {
+      const hintsConf = JSON.parse(treatment.hints);
+      const configuration = hintsConf[player.get("index").toString()];
+
+      if (!configuration) {
+        return null;
+      }
+
+      // Everytime there is a non-numeric entry in the configuration, it will become a random hint.
+      // Keep track of how many non-numeric entries there are
+      let nbRandomHints = 0;
+      // Create a copy of the possible hints, and take out each hint that is not randomly assigned to the hints
+      // (namely, assigned based on indexes set in the configuration)
+      // (this needs to be a deepCopy to not modify the possibleHints)
+      let randomPossibleHints = deepCopy(possibleHints);
+
+      // For each value of the configuration...
+      configuration.forEach((config) => {
+        // If numeric...
+        if (typeof config == "number") {
+          // ...check that there could be a hint with this index (it doesn't return undefined)
+          if (typeof possibleHints[config - 1] != "undefined") {
+            // ...push the hint that has this index
+            hints.push(possibleHints[config - 1]);
+            // and take this hint out of the hints ready to be selected randomly
+            randomPossibleHints.splice(config - 1, 1);
+          }
+        } else {
+          // Else, increase the number of random hints
+          nbRandomHints++;
+        }
+      });
+
+      // For every random hint that is needed...
+      if (nbRandomHints > 0) {
+        for (let i = 0; i < nbRandomHints; i++) {
+          // Randomly select a hint (and take it out) with popChoice
+          hints.push(popChoice(randomPossibleHints));
+        }
+      }
+    } else {
+      // Else, just allocate all the hints
+      hints = possibleHints;
+    }
+
+    // Populate the player's hints
+    let playerHints = player.get("hints");
+    playerHints[round.get("index")] = hints;
+    player.set("hints", playerHints);
+  }
+
+  return hints;
 }
